update: /embedのリファクタリング
